moves scripts around and creates example dir
