+++ example/foo
@@ -1,0 +1,1 @@
+{"top_chain_header":{"entry_type":"AgentId","entry_address":"QmQebQi1k4QpCUbEohx9NwptuXrpijt8rRXijYjzKe1fNY","entry_signature":"","link":"QmSZjwa2BJwqFbFCixomQb3RcV5NaHV3XfiWa3ecUWu97m","link_same_type":null,"timestamp":""}}
